+++ src/main/java/de/mainzelliste/paths/processor/MainzellisteClient.java
@@ -4,125 +4,110 @@ import com.sun.jersey.api.client.Client;
 import com.sun.jersey.api.client.ClientResponse;
 import com.sun.jersey.api.client.GenericType;
 import com.sun.jersey.api.client.WebResource;
-import de.mainzelliste.paths.adapters.Adapter;
-import de.mainzelliste.paths.adapters.AdapterFactory;
-import de.mainzelliste.paths.backend.Controller;
-import de.mainzelliste.paths.configuration.ConfigurationBackend;
 import de.mainzelliste.paths.configuration.Path;
-import de.pseudonymisierung.mainzelliste.client.*;
-import de.samply.common.config.Configuration;
-import de.samply.common.http.HttpConnector;
-import de.samply.common.http.HttpConnectorException;
-import org.codehaus.jettison.json.JSONException;
-import org.codehaus.jettison.json.JSONObject;
-
-import javax.ws.rs.ProcessingException;
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
+import de.pseudonymisierung.controlnumbers.ControlNumber;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
+import javax.ws.rs.ProcessingException;
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.MultivaluedHashMap;
+import javax.ws.rs.core.MultivaluedMap;
+import javax.ws.rs.core.Response;
+import org.codehaus.jettison.json.JSONException;
+import org.codehaus.jettison.json.JSONObject;
 
 public class MainzellisteClient extends AbstractProcessor {
     private final Client webClient;
-    private MainzellisteConnection mainzellisteConnection;
+    private final String mainzellisteUrl;
 
     /**
-     * Default constructor. Initializes name (see {@link #getPathName()}) and
-     * parameter map (see {@link #getParameters()}) from the given
-     * configuration.
+     * Default constructor. Initializes name (see {@link #getPathName()}) and parameter map (see
+     * {@link #getParameters()}) from the given configuration.
      *
-     * @param configuration
-     *            Configuration of this path.
+     * @param configuration Configuration of this path.
      */
     public MainzellisteClient(Path configuration) {
         super(configuration);
 
-        String mainzellisteUrl = this.getParameters().get("mainzellisteURL").getValue();
-        String mainzellisteApiKey = this.getParameters().get("mainzellisteApiKey").getValue();
-
+        String mainzellisteUrlParam = this.getParameters().get("mainzellisteURL").getValue();
+        this.mainzellisteUrl =
+            mainzellisteUrlParam + (mainzellisteUrlParam.endsWith("/") ? "" : "/");
+        // check url of maizelliste
         try {
-            HttpConnector httpConnector = Controller.getHttpConnector();
-            mainzellisteConnection = new MainzellisteConnection(mainzellisteUrl.toString(), mainzellisteApiKey,
-                    httpConnector.getHttpClient(mainzellisteUrl));
-            webClient = httpConnector.getJerseyClientForHTTPS();
+            new URI(mainzellisteUrl);
         } catch (URISyntaxException e) {
-            // logger.fatal("Invalid URI for Mainzelliste connection: " +
-            // mainzellisteUrl);
-            throw new Error("Invalid URI for Mainzelliste connection: " + mainzellisteUrl);
-            // } catch (HttpConnectorException e) {
-            // e.printStackTrace();
-            // throw new Error("Cannot connect to Mainzelliste");
+            throw new WebApplicationException("Invalid Mainzelliste Url: " + mainzellisteUrl);
         }
+
+        webClient = Client.create();
+        webClient.setFollowRedirects(false);
     }
 
     @Override
     public Map<String, Object> process(Map<String, Object> stringObjectMap) {
         Map<String, Object> output = new HashMap<>();
-        String uriString = mainzellisteConnection.getMainzellisteURI() + "patients?tokenId=" + stringObjectMap.get("KNTKT");
 
         URI idUri;
-
         try {
-            idUri = new URI(uriString);
+            idUri = new URI(mainzellisteUrl + "patients?tokenId=" + stringObjectMap.get("tokenId"));
         } catch (URISyntaxException e) {
-//            logger.error("Invalid URI for ID request: " + uriString, e);
-            throw new ProcessingException("A request used an invalid URI. See log for details.", e);
+            throw new ProcessingException("Invalid Mainzelliste uri", e);
         }
 
-        if (idUri != null) {
-            WebResource idResource = webClient.resource(idUri);
-//            logger.info("Requesting ID: POST " + idUri);
-            ClientResponse response = idResource
-                    .accept(MediaType.APPLICATION_JSON)
-                    .header("mainzellisteApiVersion", "2.0")
-                    .type(MediaType.APPLICATION_FORM_URLENCODED)
-                    .post(ClientResponse.class);
-            if (response.getStatus() >= 400) {
-                throw new WebApplicationException(
-                        Response.status(response.getStatus())
-                                .entity("Request to Mainzelliste failed with: " + response.getEntity(String.class))
-                                .build());
-            }
+        // prepare form data
+        MultivaluedMap<String, String> formData = new MultivaluedHashMap<>(
+            stringObjectMap.entrySet().stream()
+                .filter(e -> e.getValue() instanceof ControlNumber)
+                .collect(Collectors.toMap(Map.Entry::getKey,
+                    e -> createJsonFrom(e.getKey(), (ControlNumber) e.getValue())))
+        );
 
-            try {
-                List<Map<String, Object>> responseList = response.getEntity(new GenericType<List<Map<String, Object>>>() {});
-                Map<String, Object> responseMap = responseList.get(0);
-                String id = responseMap.get("idString").toString();
-                String idType = responseMap.get("idType").toString();
-                output.put("idType", idType);
-                output.put("idString", id);
-                return output;
+        WebResource idResource = webClient.resource(idUri);
+        ClientResponse response = idResource
+            .accept(MediaType.APPLICATION_JSON)
+            .header("mainzellisteApiVersion",
+                this.getParameters().get("mainzellisteApiVersion").getValue())
+            .type(MediaType.APPLICATION_FORM_URLENCODED)
+            .post(ClientResponse.class, formData);
+        if (response.getStatus() >= 400) {
+            throw new WebApplicationException(
+                Response.status(response.getStatus())
+                    .entity(
+                        "Request to Mainzelliste failed with: " + response.getEntity(String.class))
+                    .build());
+        } else if (response.getStatus() == 303) {
+            output.put("redirect", response.getHeaders().get("location"));
+            return output;
+        }
 
-//                for (Map<String, Object> thisId : responseList) {
-//                    if (this.getMainzellisteIdType(idType).equals(thisId.get("idType")))
-//                        thisResponse.put(idType, (String) thisId.get("idString"), (Boolean) thisId.get("tentative"));
-//                }
-//                logger.info(responseList.toString());
-            } catch (Throwable t) {
-//                logger.error("IdManager caught Exception...", t);
-                throw new WebApplicationException(t, Response.Status.INTERNAL_SERVER_ERROR);
-            } finally {
-                response.close();
-            }
-        } else {
-            //TODO What is to happen if the URI is null?
-            throw new WebApplicationException(Response.Status.INTERNAL_SERVER_ERROR);
+        try {
+            List<Map<String, Object>> responseList = response
+                .getEntity(new GenericType<List<Map<String, Object>>>() {
+                });
+            Map<String, Object> responseMap = responseList.get(0);
+            output.put("idType", responseMap.get("idType").toString());
+            output.put("idString", responseMap.get("idString").toString());
+            return output;
+        } catch (RuntimeException exception) {
+            throw new WebApplicationException("Invalid Mainzelliste response body", exception);
+        } finally {
+            response.close();
         }
     }
 
-    /**
-     * Get the Mainzelliste session for this instance. The session will be
-     * recreated if invalid (i.e. due to timing out).
-     *
-     * @return The Mainzelliste session.
-     * @throws MainzellisteNetworkException
-     */
-    private Session getSession() throws MainzellisteNetworkException {
-        Session mainzellisteSession = mainzellisteConnection.createSession();
-        return mainzellisteSession;
+    private String createJsonFrom(String key, ControlNumber value){
+        try {
+            JSONObject json = new JSONObject();
+            json.put("keyId", key.substring(0, key.length() - 2));
+            json.put("value", value.toBitString());
+            return json.toString();
+        } catch (JSONException e) {
+            return "";
+        }
     }
 }