+++ changelog.md
@@ -2,6 +2,7 @@
 ### Added
 - docker support - see readme.md for more information about docker
 - Support for CORS Requests. The CORS access can be specified by using PATHS_ALLOWED_ORIGINS and PATHS_ALLOWED_CORS_HEADERS environment variables.
+- Support redirect response of mainzelliste
 ### Changed
 - Default config file name from pathsDKTK.xml to paths.xml
 - Updated dependency de.samply.common-config from 2.3.0 to 3.1.0